# Remove call to time.Now() on worker thread when handling record reqs